test that newly created GPT tables overwrite/preserve the right areas
When working with a previously uninitialized disk, it's important that 
certain areas of the header sector and partition array are cleared, 
whereas any actual data areas should be left untouched. Test that this 
works properly for both 512-byte and 4096-byte sector sizes.

Add another example to the mod docs that demonstrates how to create a 
partition table completely from scratch, add a few partitions, and write 
it to a memory buffer (could just as easily could have been a file).