make GptDisk generic over DiskDeviceObject lifetimes
Normally the rust compiler is able to infer lifetimes, and so this 
should not often require additional annotations when using GptDisk.

This modification allows the use of DiskDeviceObjects in a GptDisk whose 
lifetime is something other than 'static, e.g., a local variable whose 
ownership needs to be returned once GptDisk is dropped (or when write() 
is called).