clarify that ProtectiveMBR::with_lb_size needs the size of the partition
Previously, the documentation said that it needed the size of the disk 
(in LBs). However, the MBR partition entry that is constructed actually 
uses this value without adjustment as the size of the partition. Since 
the protective partition always begins at LBA 1 (the second sector), 
ProtectiveMBR::with_lb_size should be passed the size of the disk (in 
LBs) minus 1.