Support for arbitrary devices & other various enhancements (#59)
* add ability to convert u64 into LogicalBlockSize

* add ability to convert from partition type name to partition type

Useful if partitions will be constructed from program input (which 
favors human-friendly names in addition to GUIDs) rather than code 
(which can use constants for GUIDs)

* add ability to work with arbitrary Read + Write + Seek + Debug objects

GptDisk, Partition, and ProtectiveMBR can now work with anything that 
implements the required traits, instead of just a File.

Backwards-compatible from a typing perspective, except for 
GptDisk::write, which now returns the DiskDeviceObject trait object 
(e.g., a boxed File) instead of the File object directly.

* bugfix: GptDisk::write must write backup header partition array

Otherwise the backup header will be valid (have a valid CRC), but 
potentially point to different partition array data (at the tail of the 
disk).

* test that if we write what we read then it will match

This test exposed the bug fixed in the prior commit where the backup 
header did not match before the bugfix.

Also test new capabilities using the abstract DiskDeviceObject trait 
object using a memory cursor, and compare to file-based results.

* add GptConfig::create_from_device so that disk uuid can be specified

* the unused part of the sector with the header should be zero-filled

See page 119 in 
https://uefi.org/sites/default/files/resources/UEFI_Spec_2_8_final.pdf

* More UEFI compliant headers when creating one from scratch

Partition tables created from scratch should have room for more entries 
than just the number of partitions we initially have (e.g., t least 
128).

Also, the default first usable sector needs to consider the block size, 
rather than just being hard coded to 34.

The last usable sector calculation should be based on the number of 
logical blocks required to store the partition array, rather than the 
first usable sector.

* write zeros to clear unused entries in partition array

Now that the header contains the number of total entries in the
partition array (and not just the number populated), it's important that
any remaining entries are zero-filled, so that any previously existing
partition entries on disk are removed.

* test that newly created GPT tables overwrite/preserve the right areas

When working with a previously uninitialized disk, it's important that 
certain areas of the header sector and partition array are cleared, 
whereas any actual data areas should be left untouched. Test that this 
works properly for both 512-byte and 4096-byte sector sizes.

Add another example to the mod docs that demonstrates how to create a 
partition table completely from scratch, add a few partitions, and write 
it to a memory buffer (could just as easily could have been a file).

* make GptDisk generic over DiskDeviceObject lifetimes

Normally the rust compiler is able to infer lifetimes, and so this 
should not often require additional annotations when using GptDisk.

This modification allows the use of DiskDeviceObjects in a GptDisk whose 
lifetime is something other than 'static, e.g., a local variable whose 
ownership needs to be returned once GptDisk is dropped (or when write() 
is called).

* expose MBR PartRecord entries and fields as pub

Allows this crate to be useful for editing MBR partition tables in 
general (not just to generate protective MBRs)

* clarify that ProtectiveMBR::with_lb_size needs the size of the partition

Previously, the documentation said that it needed the size of the disk 
(in LBs). However, the MBR partition entry that is constructed actually 
uses this value without adjustment as the size of the partition. Since 
the protective partition always begins at LBA 1 (the second sector), 
ProtectiveMBR::with_lb_size should be passed the size of the disk (in 
LBs) minus 1.

* Bump MSRV to 1.34 (needed for TryFrom trait)

* clippy cleanup for rust 1.34.0

Now that MSRV has bumped to 1.34.0, some clippy warnings got triggered. 
Fix these.

* avoid panics when getting/setting partitions in ProtectiveMBR