find_free_sectors: fix allocation of one-sector partitions at the start
The fix which special-cases the first usable block introduced a bug
when the first partition is one block long. The issue is that the
values in p[0] typically represent the last in-use block. That fix,
however, changed p[0] to mean the last *free* block as long as p[0]
was the first usable block. This is not correct, and causes problems
if the first partition is exactly the first block.

Assuming 34 is the first usable block and 94 is the last, we'd see these
values in the mapping:

(34 34)
(34, 94)

Because the second iteration also starts with 34, the partition which
exists entirely in block 34 gets overwritten.

Furthermore, there appears to be an issue with one-block partitions
"fitting" in 0-length free space, so I added a filter to remove
zero-length sectors.